requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain-groq
faiss-cpu
pandas
matplotlib
sentence-transformers